+++ changelog.md
@@ -3,6 +3,17 @@
 Changelog template based on:
 https://keepachangelog.com/en/1.1.0/
 
+## [2.0.0] - 2025-06-05
+
+### Added
+
+- Launch file for different (debug) cases
+
+### Changed
+
+- Config - Everything config related is now part of `conf` call, config file works the same, but the commands have changed: path, file, to, nto, ignore, heed, here, away are now used instead commands: pfp, pfc, (timeout was not configurable), ignore, unignore, swd, uwd.
+- Default timeout for Task Runner was removed, timeout is configurable now as conf command or as a temporary flag. Sustain works as previously.
+
 ## [1.1.2] - 2024-10-23
 
 ### Added